// packages/core/tests/fixtures/zenstack/schema.ts
export class SchemaType implements SchemaDef {
    provider = {
        type: "sqlite"
    } as const;
    models = {
        User: {
            name: "User",
            fields: {
                id: {
                    name: "id",
                    type: "String",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("cuid") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("cuid") as FieldDefault
                },
                email: {
                    name: "email",
                    type: "String",
                    unique: true,
                    attributes: [{ name: "@unique" }, { name: "@email" }] as readonly AttributeApplication[]
                },
// ...
                address: {
                    name: "address",
                    type: "Address",
                    optional: true,
                    attributes: [{ name: "@json" }] as readonly AttributeApplication[]
                },
                settings: {
                    name: "settings",
                    type: "Json",
                    optional: true
                },
                balance: {
                    name: "balance",
                    type: "Decimal",
                    attributes: [{ name: "@default", args: [{ name: "value", value: ExpressionUtils.literal(0) }] }] as readonly AttributeApplication[],
                    default: 0 as FieldDefault
                },
                secret: {
                    name: "secret",
                    type: "Bytes",
                    optional: true,
                    omit: true,
                    attributes: [{ name: "@omit" }] as readonly AttributeApplication[]
                },
                postCount: {
                    name: "postCount",
                    type: "Int",
                    attributes: [{ name: "@computed" }] as readonly AttributeApplication[],
                    computed: true
                },
                assets: {
                    name: "assets",
                    type: "Asset",
                    array: true,
                    relation: { opposite: "owner" }
                }
            },
// ...
        },
// ...
        Post: {
            name: "Post",
            fields: {
// ...
                author: {
                    name: "author",
                    type: "User",
                    attributes: [{ name: "@relation", args: [{ name: "fields", value: ExpressionUtils.array("String", [ExpressionUtils.field("authorId")]) }, { name: "references", value: ExpressionUtils.array("String", [ExpressionUtils.field("id")]) }, { name: "onDelete", value: ExpressionUtils.literal("Cascade") }] }] as readonly AttributeApplication[],
                    relation: { opposite: "posts", fields: ["authorId"], references: ["id"], onDelete: "Cascade", hasDefault: true }
                },
// ...
            },
// ...
        },
        Attachment: {
            name: "Attachment",
            fields: {
                id: {
                    name: "id",
                    type: "String",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("cuid") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("cuid") as FieldDefault
                },
                post: {
                    name: "post",
                    type: "Post",
                    attributes: [{ name: "@relation", args: [{ name: "fields", value: ExpressionUtils.array("String", [ExpressionUtils.field("postId")]) }, { name: "references", value: ExpressionUtils.array("String", [ExpressionUtils.field("id")]) }, { name: "onDelete", value: ExpressionUtils.literal("Cascade") }] }] as readonly AttributeApplication[],
                    relation: { opposite: "attachments", fields: ["postId"], references: ["id"], onDelete: "Cascade" }
                },
// ...
            },
// ...
        },
// ...
        Video: {
            name: "Video",
            baseModel: "Asset",
            fields: {
                id: {
                    name: "id",
                    type: "String",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("cuid") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("cuid") as FieldDefault
                },
                createdAt: {
                    name: "createdAt",
                    type: "DateTime",
                    originModel: "Asset",
                    attributes: [{ name: "@default", args: [{ name: "value", value: ExpressionUtils.call("now") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("now") as FieldDefault
                },
                owner: {
                    name: "owner",
                    type: "User",
                    originModel: "Asset",
                    attributes: [{ name: "@relation", args: [{ name: "fields", value: ExpressionUtils.array("String", [ExpressionUtils.field("ownerId")]) }, { name: "references", value: ExpressionUtils.array("String", [ExpressionUtils.field("id")]) }, { name: "onDelete", value: ExpressionUtils.literal("Cascade") }] }] as readonly AttributeApplication[],
                    relation: { opposite: "assets", fields: ["ownerId"], references: ["id"], onDelete: "Cascade", hasDefault: true }
                },
                ownerId: {
                    name: "ownerId",
                    type: "String",
                    originModel: "Asset",
                    attributes: [{ name: "@default", args: [{ name: "value", value: ExpressionUtils.member(ExpressionUtils.call("auth"), ["id"]) }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.member(ExpressionUtils.call("auth"), ["id"]) as FieldDefault,
                    foreignKeyFor: [
                        "owner"
                    ] as readonly string[]
                },
                kind: {
                    name: "kind",
                    type: "String",
                    originModel: "Asset",
                    isDiscriminator: true
                },
                preview: {
                    name: "preview",
                    type: "String",
                    optional: true,
                    originModel: "Asset",
                    attributes: [{ name: "@file", args: [{ name: "cleanup", value: ExpressionUtils.literal(true) }] }] as readonly AttributeApplication[]
                },
                duration: {
                    name: "duration",
                    type: "Int"
                },
                source: {
                    name: "source",
                    type: "String",
                    optional: true,
                    attributes: [{ name: "@file", args: [{ name: "accept", value: ExpressionUtils.array("String", [ExpressionUtils.literal("video/*")]) }, { name: "cleanup", value: ExpressionUtils.literal(true) }] }] as readonly AttributeApplication[]
                }
            },
            attributes: [
                { name: "@@allow", args: [{ name: "operation", value: ExpressionUtils.literal("all") }, { name: "condition", value: ExpressionUtils.binary(ExpressionUtils.call("auth"), "==", ExpressionUtils.field("owner")) }] }
            ] as readonly AttributeApplication[],
            idFields: ["id"],
            uniqueFields: {
                id: { type: "String" }
            }
        },
// ...
    } as const;
    typeDefs = {
        Address: {
            name: "Address",
            fields: {
                street: {
                    name: "street",
                    type: "String"
                },
                city: {
                    name: "city",
                    type: "String"
                },
                zip: {
                    name: "zip",
                    type: "String",
                    optional: true
                }
            }
        }
    } as const;
    enums = {
        Role: {
            name: "Role",
            values: {
                USER: "USER",
                ADMIN: "ADMIN"
            }
        }
    } as const;
    authType = "User" as const;
    plugins = {};
}
export const schema = new SchemaType();